# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/notify/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notify"
                ],
                "summary": "创建群通知",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ParamCreateNotify"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败 2009:权限不足 7001:非群主 7003:非群成员",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ParamGroupNotify"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/notify/delete": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notify"
                ],
                "summary": "获取群通知",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "群通知的 ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "群 ID",
                        "name": "relation_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败 2009:权限不足 7001:非群主 7003:非群成员",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
        "/api/notify/get": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notify"
                ],
                "summary": "更具群ID获取群通知",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "relation_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败 2009:权限不足 7003:非群成员",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ParamGetNotifyByID"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/notify/update": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notify"
                ],
                "summary": "更新群通知",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ParamUpdateNotify"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败 2009:权限不足 7001:非群主 7003:非群成员",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
...
        "/api/user/deleteUser": {
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2001:用户不存在 2007:身份不存在 2008:身份验证失败",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
        "/api/user/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "用户登录信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ParamLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数错误 1003:系统错误 2001:用户不存在",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ParamLogin"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/user/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "用户注册信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ParamRegister"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2004:邮箱验证码校验失败 2006:邮箱已经注册",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ParamRegister"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/user/update/email": {
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "更新用户邮箱",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "新邮箱和验证码",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ParamUpdateUserEmail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2001:用户不存在 2004:邮箱验证码校验失败 2005:邮箱相同 2006:邮箱已经注册 2007:身份不存在 2008:身份验证失败",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
        "/api/user/update/pwd": {
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "更新用户密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "验证码和新密码",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ParamUpdateUserPassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2001:用户不存在 2004:邮箱验证码校验失败 2007:身份不存在 2008:身份验证失败",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.State": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码，0：成功；否则失败",
                    "type": "integer"
                },
                "data": {
                    "description": "数据，失败时返回空"
                },
                "msg": {
                    "description": "状态的具体描述",
                    "type": "string"
                }
            }
        },
        "common.Token": {
            "type": "object",
            "properties": {
                "expire_at": {
                    "description": "token 过期时间",
                    "type": "string"
                },
                "token": {
                    "description": "token",
                    "type": "string"
                }
            }
        },
        "model.MsgExtend": {
            "type": "object",
            "properties": {
                "remind": {
                    "description": "@ 的描述信息",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Remind"
                    }
                }
            }
        },
        "model.Remind": {
            "type": "object",
            "required": [
                "account_id",
                "idx"
            ],
            "properties": {
                "account_id": {
                    "description": "被 @ 的账号 ID",
                    "type": "integer",
                    "minimum": 1
                },
                "idx": {
                    "description": "第几个 @",
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "reply.ParamGetNotifyByID": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/reply.ParamGroupNotify"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "reply.ParamGroupNotify": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "integer"
                },
                "create_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "msg_content": {
                    "type": "string"
                },
                "msg_extend": {
                    "$ref": "#/definitions/model.MsgExtend"
                },
                "read_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "relation_id": {
                    "type": "integer"
                }
            }
        },
...
        "reply.ParamLogin": {
            "type": "object",
            "properties": {
                "param_user_info": {
                    "description": "用户信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/reply.ParamUserInfo"
                        }
                    ]
                },
                "token": {
                    "description": "用户令牌",
                    "allOf": [
                        {
                            "$ref": "#/definitions/common.Token"
                        }
                    ]
                }
            }
        },
...
        "reply.ParamRegister": {
            "type": "object",
            "properties": {
                "param_user_info": {
                    "description": "用户信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/reply.ParamUserInfo"
                        }
                    ]
                },
                "token": {
                    "description": "用户令牌",
                    "allOf": [
                        {
                            "$ref": "#/definitions/common.Token"
                        }
                    ]
                }
            }
        },
...
        "reply.ParamUserInfo": {
            "type": "object",
            "properties": {
                "create_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "id": {
                    "description": "user id",
                    "type": "integer"
                }
            }
        },
...
        "request.ParamCreateNotify": {
            "type": "object",
            "required": [
                "msg_content",
                "relation_id"
            ],
            "properties": {
                "msg_content": {
                    "type": "string"
                },
                "msg_extend": {
                    "$ref": "#/definitions/model.MsgExtend"
                },
                "relation_id": {
                    "type": "integer"
                }
            }
        },
...
        "request.ParamLogin": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "邮箱",
                    "type": "string",
                    "maxLength": 50
                },
                "password": {
                    "description": "验证码",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 6
                }
            }
        },
...
        "request.ParamRegister": {
            "type": "object",
            "required": [
                "code",
                "email",
                "password"
            ],
            "properties": {
                "code": {
                    "description": "验证码",
                    "type": "string",
                    "maxLength": 6,
                    "minLength": 6
                },
                "email": {
                    "description": "邮箱（required 要求该字段必须有值）",
                    "type": "string"
                },
                "password": {
                    "description": "密码（该字段是必须的，并且长度介于6——50之间）",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 6
                }
            }
        },
...
        "request.ParamUpdateNotify": {
            "type": "object",
            "required": [
                "id",
                "msg_content",
                "relation_id"
            ],
            "properties": {
                "id": {
                    "description": "群通知 ID",
                    "type": "integer"
                },
                "msg_content": {
                    "type": "string"
                },
                "msg_extend": {
                    "$ref": "#/definitions/model.MsgExtend"
                },
                "relation_id": {
                    "type": "integer"
                }
            }
        },
...
        "request.ParamUpdateUserEmail": {
            "type": "object",
            "required": [
                "code",
                "email"
            ],
            "properties": {
                "code": {
                    "description": "验证码",
                    "type": "string",
                    "maxLength": 6,
                    "minLength": 6
                },
                "email": {
                    "description": "邮箱",
                    "type": "string",
                    "maxLength": 50
                }
            }
        },
        "request.ParamUpdateUserPassword": {
            "type": "object",
            "required": [
                "code",
                "newPassword"
            ],
            "properties": {
                "code": {
                    "description": "验证码",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 6
                },
                "newPassword": {
                    "description": "新的密码",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 6
                }
            }
        },
...
    }
}